requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pandas>=2.2.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
pytest>=8.0.0
requests>=2.32.0
jinja2>=3.1.0
pdfkit>=1.0.0
python-dotenv>=1.0.0
python-dateutil>=2.9.0
reportlab>=4.0.0
openpyxl>=3.1.0
fastapi
uvicorn
pandas